fix ip address lost after stopping the daemon
The daemon was not unbinding the GPU queues from the rx queues on stop.
To unbind the daemon must:

```
close(dma_buf_pages_fd);
ethtool --set-priv-flags enable-header-split on // make sure header split is on
ethtool --set-priv-flags enable-header-split off // turn it off, triggering a driver reset.
```

Tested: started and stopped the daemon many times, I don't see the IP
address lost. Before this fix, IP address was lost 100% of the time on
eth1 on stopping the daemon.

Change-Id: I329a984741eac1c8b124c8768f456d51eb266a67
Reviewed-on: https://team-review.git.corp.google.com/c/kernel-net-team/tcpgpudmarxd/+/1931075
Reviewed-by: Che Lin <chechenglin@google.com>
Reviewed-by: Art Zhu <artzhu@google.com>