add .clang-format and format all files
Added a .clang-format with Google style. One hint is that if we want to
make all pointers shift to left, we should make sure that
DerivePointerAlignment in .clang-format is false and PointerAlignment is
left. Otherwise, clang will make everything consistent by following the
majority alignment in each file.

Vscode user now with clang-format tool will automatically use the style
defined in .clang-format.

Change-Id: I026d7ae9cadf6e6ff0a329a04e3545be9b230c41
Reviewed-on: https://team-review.git.corp.google.com/c/kernel-net-team/tcpgpudmarxd/+/1978246
Reviewed-by: Che Lin <chechenglin@google.com>