Revert "Adding python3 to docker build file, this"
This reverts commit d0e6cbf559fd7ea442b5f5c66c54efcd0ea32a94.

Reason for revert: 

This change has been moved to pending branch. In order to better roll out unified A3/A3+ unified RxDM image, the HEAD of dev repo will be pinged to the latest image that has been qualified and released to customers once we release the unified image. Currently we don't have resource to do any large scale qualification on A3 so we can't roll up to HEAD with this changes. 

Change-Id: I8f942588f0b691b0f5bb5318a8a163b17132dd9e
Reviewed-on: https://team-review.git.corp.google.com/c/kernel-net-team/tcpgpudmarxd/+/2262495
Reviewed-by: Kaikai Zhu <kaikaiz@google.com>