Revert "Add tuning persistance to a3-tunnings in rxdm"
This reverts commit 4a3c5545ba95ba3279645f0165afdeb2f745506a.

Reason for revert: 

This change has been moved to pending branch. In order to better roll out unified A3/A3+ unified RxDM image, the HEAD of dev repo will be pinged to the latest image that has been qualified and released to customers once we release the unified image. Currently we don't have resource to do any large scale qualification on A3 so we can't roll up to HEAD with this changes. 

Change-Id: I3061dcb037905116e47e3edc85bb9cd2b458863f
Reviewed-on: https://team-review.git.corp.google.com/c/kernel-net-team/tcpgpudmarxd/+/2262496
Reviewed-by: Kaikai Zhu <kaikaiz@google.com>